third_party/WebKit/LayoutTests/typedcssom/cssTranslation.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

var zeroLength = new CSSUnitValue(0, "px");
var decimalLength = new CSSUnitValue(1.1, "px");
var negativeLength = new CSSUnitValue(-2.2, "em");
var tenPercent = new CSSUnitValue(10, "percent");

var testParams = [
  // 2D CSSTranslation Transform Components
  {
    input: new CSSTranslation(zeroLength, zeroLength),
    x: zeroLength, y: zeroLength,
    is2D: true,
    cssText: "translate(0px, 0px)"
  },
  {
    input: new CSSTranslation(decimalLength, negativeLength),
    x: decimalLength, y: negativeLength,
    is2D: true,
    cssText: "translate(1.1px, -2.2em)"
  },
  {
    input: new CSSTranslation(tenPercent, zeroLength),
    x: tenPercent, y: zeroLength,
    is2D: true,
    cssText: "translate(10%, 0px)"
  },

  // 3D CSSTranslation Transform Components
  {
    input: new CSSTranslation(zeroLength, zeroLength, zeroLength),
    x: zeroLength, y: zeroLength, z: zeroLength,
    is2D: false,
    cssText: "translate3d(0px, 0px, 0px)"
  },
  {
    input: new CSSTranslation(zeroLength, decimalLength, negativeLength),
    x: zeroLength, y: decimalLength, z: negativeLength,
    is2D: false,
    cssText: "translate3d(0px, 1.1px, -2.2em)"
  },
  {
    input: new CSSTranslation(tenPercent, decimalLength, zeroLength),
    x: tenPercent, y: decimalLength, z: zeroLength,
    is2D: false,
    cssText: "translate3d(10%, 1.1px, 0px)"
  },
];

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.x, params.x);
    assert_equals(params.input.y, params.y);
    if (params.is2D) {
      assert_equals(params.input.z, null);
    } else {
      assert_equals(params.input.z, params.z);
    }
  }, "x, y, and z values are correct for " + params.cssText);
}

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.is2D(), params.is2D);
  }, "is2D value is correct for " + params.cssText);
}

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.toString(), params.cssText);
  }, "toString value is correct for " + params.cssText);
}

test(() => {
  let translation = new CSSTranslation(
      new CSSUnitValue(10, 'percent'),
      new CSSUnitValue(20, 'percent'));
  assert_equals(translation.x.value, 10);
  assert_equals(translation.x.unit, 'percent');
  assert_equals(translation.y.value, 20);
  assert_equals(translation.y.unit, 'percent');
  assert_equals(translation.z, null);

  let translation3d = new CSSTranslation(
      new CSSUnitValue(30, 'percent'),
      new CSSUnitValue(40, 'percent'),
      zeroLength);
  assert_equals(translation3d.x.value, 30);
  assert_equals(translation3d.x.unit, 'percent');
  assert_equals(translation3d.y.value, 40);
  assert_equals(translation3d.y.unit, 'percent');
  assert_equals(translation3d.z.value, 0);
}, "Constructor accepts percent for x and y");

test(() => {
  assert_throws(new TypeError(), () => {
    new CSSTranslation(new CSSUnitValue(10, 'deg'), zeroLength);
  });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(zeroLength, new CSSUnitValue(10, 'deg'));
  });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(new CSSUnitValue(10, 'deg'), zeroLength, zeroLength);
  });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(zeroLength, new CSSUnitValue(10, 'deg'), zeroLength);
  });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(zeroLength, zeroLength, new CSSUnitValue(10, 'deg'));
  });
}, "Constructor throws when invalid numeric values are given to each argument");

test(() => {
  assert_throws(new TypeError(), () => {
    new CSSTranslation(zeroLength, zeroLength, tenPercent);
  });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(tenPercent, tenPercent, tenPercent);
  });
}, "Constructor throws when z argument contains percent.");

test(() => {
  assert_throws(new TypeError(), () => { new CSSTranslation(); });
  assert_throws(new TypeError(), () => { new CSSTranslation(zeroLength); });
}, "Invalid number of arguments to constructor throws an exception.");

</script>
